# Optimize rendering and processing latencies for Time Lapse View Hierarchies